rename: restaurant 관련 파일명 변경
