+++ src/App.tsx
@@ -1,4 +1,4 @@
-import FilterableMenuTable from './components/FilterableProductTable';
+import FilterableMenuTable from './components/FilterableRestaurantsTable';
 
 // - 사용자는 식당 이름을 입력하여 이름이 (부분)일치하는 식당 목록을 골라 볼 수 있다.
 // - 사용자는 식당 종류 버튼을 눌러서 종류가 일치하는 식당 목록만 골라 볼 수 있다.
