+++ src/components/RestaurantsRow.tsx
@@ -1,0 +1,14 @@
+import RestaurantItem from './RestaurantItem';
+import { restaurants } from '../../restaurants.json';
+
+function RestaurantsRow() {
+  return (
+    <tbody>
+      {restaurants.map((restaurant) => (
+        <RestaurantItem key={restaurant.name} restaurant={restaurant} />
+      ))}
+    </tbody>
+  );
+}
+
+export default RestaurantsRow;
